requirements.txt
flask
gunicorn
line-bot-sdk>=3.0.0,<4
requests
pandas
matplotlib
transformers
torch
accelerate
safetensors
huggingface_hub
sentencepiece